requirements.txt
flask==3.0.3
g4f== 0.4.6.1
torch
torchvision
pillow==10.4.0
flask-cors==5.0.0
